add check to skip referers from this run
