+++ cyberdrop_dl/scraper/crawler.py
@@ -89,7 +89,7 @@ class Crawler(ABC):
         
         check_referer = False
         if self.manager.config_manager.settings_data['Download_Options']['skip_referer_seen_before']:
-            check_referer = await self.manager.db_manager.history_table.check_referer(scrape_item.url)
+            check_referer = await self.manager.db_manager.temp_referer_table.check_referer(scrape_item.url)
 
         if check_referer:
             await log(f"Skipping {url} as referer has been seen before", 10)
