+++ cyberdrop_dl/scraper/scraper.py
@@ -454,7 +454,7 @@ class ScrapeMapper:
                 posible_referer = scrape_item.parents[-1]
                 check_referer = False
                 if self.manager.config_manager.settings_data['Download_Options']['skip_referer_seen_before']:
-                    check_referer = await self.manager.db_manager.history_table.check_referer(posible_referer)
+                    check_referer = await self.manager.db_manager.temp_referer_table.check_referer(posible_referer)
 
                 if check_referer:
                     await log(f"Skipping {scrape_item.url} as referer has been seen before", 10)
